Fix corrupt dedup output for inputs with quotes
When running dedup on a pairsum files that includes quotes in the QUAL field, the result would be a corrupt file. The to_csv method would quote the entire column, and would also escape the quote with a second quote. This results in a file that has QUAL and SEQ of different lengths. To fix this, we ask DataFrame.to_csv to never quote the output.